+++ src/ContainerView.h
@@ -29,7 +29,7 @@ private:
 
 	bool transitioning_;
 	float interval_;
-	int step_;
+	float step_;
 	void move_left();
 	void move_right();
